v4.1.x opal/verbs: disable calling of ibv_fork_init() by default
Currently function ibv_fork_init() was called by default during the
initialization of btl/openib component, which enabled rdma-core's
fork support. This patch disables it by settining the default value of
opal_common_verbs_want_fork_support to 0.

The reason for this change is because of the high cost of rdma-core's
fork support when RDMAV_HUGEPAGE_SAFE is set.

If the environment variable RDMAV_HUGEPAGE_SAFE is set, rdma-core's
fork support became very expensive because it try to get page size from
/proc/<pid>/smaps.

Note this would affect an application that does not call fork().

Given that opal already registered an atfork callback function that
warns user about the risk of using fork(). We believe it should be user's
choice to enable rdma-core's fork support by setting the environment
variable IBV_FORK_SAFE.

In the master branch, the btl/openib component was removed completely
and there is no more a call to ibv_for_init(). Therefore this patch
applies to v4.1.x branch directly.

Signed-off-by: Wei Zhang <wzam@amazon.com>